Add detector tests back
