+++ gcs/get_gcs_test.go
@@ -238,43 +238,3 @@ func TestGetter_Url(t *testing.T) {
 		})
 	}
 }
-
-func TestGetter_Detect(t *testing.T) {
-	cases := []struct {
-		Input  string
-		Output string
-	}{
-		{
-			"www.googleapis.com/storage/v1/bucket/foo",
-			"https://www.googleapis.com/storage/v1/bucket/foo",
-		},
-		{
-			"www.googleapis.com/storage/v1/bucket/foo/bar",
-			"https://www.googleapis.com/storage/v1/bucket/foo/bar",
-		},
-		{
-			"www.googleapis.com/storage/v1/foo/bar.baz",
-			"https://www.googleapis.com/storage/v1/foo/bar.baz",
-		},
-	}
-
-	pwd := "/pwd"
-	f := new(Getter)
-	for i, tc := range cases {
-		req := &getter.Request{
-			Src: tc.Input,
-			Pwd: pwd,
-		}
-		ok, err := getter.Detect(req, f)
-		if err != nil {
-			t.Fatalf("err: %s", err)
-		}
-		if !ok {
-			t.Fatal("not ok")
-		}
-
-		if req.Src != tc.Output {
-			t.Fatalf("%d: bad: %#v", i, req.Src)
-		}
-	}
-}
