+++ s3/get_s3_test.go
@@ -312,86 +312,3 @@ func TestGetter_Url(t *testing.T) {
 		})
 	}
 }
-
-func TestGetter_Detect(t *testing.T) {
-	cases := []struct {
-		Input  string
-		Output string
-	}{
-		// Virtual hosted style
-		{
-			"bucket.s3.amazonaws.com/foo",
-			"https://s3.amazonaws.com/bucket/foo",
-		},
-		{
-			"bucket.s3.amazonaws.com/foo/bar",
-			"https://s3.amazonaws.com/bucket/foo/bar",
-		},
-		{
-			"bucket.s3.amazonaws.com/foo/bar.baz",
-			"https://s3.amazonaws.com/bucket/foo/bar.baz",
-		},
-		{
-			"bucket.s3-eu-west-1.amazonaws.com/foo",
-			"https://s3-eu-west-1.amazonaws.com/bucket/foo",
-		},
-		{
-			"bucket.s3-eu-west-1.amazonaws.com/foo/bar",
-			"https://s3-eu-west-1.amazonaws.com/bucket/foo/bar",
-		},
-		{
-			"bucket.s3-eu-west-1.amazonaws.com/foo/bar.baz",
-			"https://s3-eu-west-1.amazonaws.com/bucket/foo/bar.baz",
-		},
-		// Path style
-		{
-			"s3.amazonaws.com/bucket/foo",
-			"https://s3.amazonaws.com/bucket/foo",
-		},
-		{
-			"s3.amazonaws.com/bucket/foo/bar",
-			"https://s3.amazonaws.com/bucket/foo/bar",
-		},
-		{
-			"s3.amazonaws.com/bucket/foo/bar.baz",
-			"https://s3.amazonaws.com/bucket/foo/bar.baz",
-		},
-		{
-			"s3-eu-west-1.amazonaws.com/bucket/foo",
-			"https://s3-eu-west-1.amazonaws.com/bucket/foo",
-		},
-		{
-			"s3-eu-west-1.amazonaws.com/bucket/foo/bar",
-			"https://s3-eu-west-1.amazonaws.com/bucket/foo/bar",
-		},
-		{
-			"s3-eu-west-1.amazonaws.com/bucket/foo/bar.baz",
-			"https://s3-eu-west-1.amazonaws.com/bucket/foo/bar.baz",
-		},
-		// Misc tests
-		{
-			"s3-eu-west-1.amazonaws.com/bucket/foo/bar.baz?version=1234",
-			"https://s3-eu-west-1.amazonaws.com/bucket/foo/bar.baz?version=1234",
-		},
-	}
-
-	pwd := "/pwd"
-	f := new(Getter)
-	for i, tc := range cases {
-		req := &getter.Request{
-			Src: tc.Input,
-			Pwd: pwd,
-		}
-		ok, err := getter.Detect(req, f)
-		if err != nil {
-			t.Fatalf("err: %s", err)
-		}
-		if !ok {
-			t.Fatal("not ok")
-		}
-
-		if req.Src != tc.Output {
-			t.Fatalf("%d: bad: %#v", i, req.Src)
-		}
-	}
-}
